Convert tests to use Verify.ImageMagick for QR-bills verification
Using Assert.Equal on byte arrays is fine as long as the tests pass. As soon as one test fails, it becomes unusable because of its useless message:
> Xunit.Sdk.EqualException
> Assert.Equal() Failure
> Expected: Byte[] [37, 80, 68, 70, 45, ...]
> Actual:   Byte[] [37, 80, 68, 70, 45, ...]

Something has changed in the way compression works in .NET Core 2 and .NET Core 3 and all PDF tests were failing because of the byte-by-byte comparison of PDF files.

Using Verify.ImageMagick instead is much better as it performs a real image comparison (with ImageMagick under the hood).